refactor(ref): add concept of `ref`s to each submodule
gen-api now creates a constructor for all modules other than the
Connect module to accept a `conn` and `attrs` that are used to make
an internal reference to the remote object of this type. Methods
attached to the prototype of each submodule now use that ref in
their rpc calls, rather than requiring the user pass one in.